multi-root: improve display of path for 'WorkspaceInputDialog'
Fixes: https://github.com/eclipse-theia/theia/issues/7309

Improved the display of the multi-root workspace. Now it displays the
absolute path of the workspace root aswell.
Previously it only used to display the parent-folder name or path.

With this update, the ambiguity for creating a new file/folder in a
multi-root wont be ambiguous. If two roots have the same name, with the
absolute path it would be clear as to which root is selected for a new
file/folder.

Signed-off-by: Anas Shahid <muhammad.shahid@ericsson.com>